src/ESPiLight.cpp: Interrupts are unsigned.

+++ src/ESPiLight.cpp
@@ -54,7 +54,7 @@ void ESPiLight::initReceiver(byte inputPin) {
   enableReceiver();
 
   if (interrupt >= 0) {
-    attachInterrupt(interrupt, interruptHandler, CHANGE);
+    attachInterrupt(static_cast<uint8_t>(interrupt), interruptHandler, CHANGE);
   }
 }
